+++ src/algorithms.rs
@@ -6,3 +6,4 @@ pub(crate) mod generate;
 pub(crate) mod oaep;
 pub(crate) mod pkcs1v15;
 pub(crate) mod pss;
+pub(crate) mod rsa;
